--- a/src/lofreq/Makefile.am
+++ b/src/lofreq/Makefile.am
@@ -41,4 +41,4 @@ LIBS_for_htslib = -lhts
 endif
 
 # note: order matters
-lofreq_LDADD = $(LIBS_for_htslib) ../cdflib90/libcdf.a -lgsl -lcblas -lm
+lofreq_LDADD = $(LIBS_for_htslib) ../cdflib90/libcdf.a -l:libgsl.a -lcblas -lm