--- a/src/lofreq/snpcaller.c
+++ b/src/lofreq/snpcaller.c
@@ -1102,13 +1102,14 @@ snpcaller(long double *snp_pvalues,
         goto free_and_exit;
     }
 
-    if (num_err_probs > 1000) {
-        double mu = 0;
+    // Only approximate if sufficient data available
+    if (num_err_probs > 100) {
+        long double mu = 0;
         for (int i = 0; i < num_err_probs; ++i) {
             mu += err_probs[i];
         }
         const long double poibin_approximation = 1 - gsl_cdf_poisson_P(max_noncons_count - 1, mu);
-        if (poibin_approximation > sig_level + 0.01) {
+        if (poibin_approximation > sig_level + 0.05) {
             goto free_and_exit;
         }
     }
@@ -1131,7 +1132,6 @@ snpcaller(long double *snp_pvalues,
 #endif
         goto free_and_exit;
     }
-
 
     /* report p-value for each non-consensus base
      */