Hello! 

This is a PR for an additional heuristic in the LoFreq algorithm. Many times, the Poisson Binomial calculation either yields numbers that are extremely high p-values extremely low p-values. Therefore, we can use an approximation to cut out performing an exact calculation when the p-value is sufficiently above the significance level. By specifying some threshold above the significance level, we can be extremely confident that the approximation  shortcut will never skip exact calculation on columns which have significant p-values. 

We use a [Poisson approximation to the binomial distribution](https://www.jstor.org/stable/2237582) where the mean is the sum of the probabilities and the variable remains the same (most frequent variant count). There are indeed other approximations that we plan on trying, such as the [refined normal approximation](https://www.emis.de/journals/HOA/IJMMS/Volume2005_5/728.pdf) (see section 3 of [this](https://www.sciencedirect.com/science/article/pii/S0167947312003568) pub for a brief explanation of the approximations). To calculate the Poisson CDF, we use the [GNU scientific library](https://www.gnu.org/software/gsl/doc/html/intro.html). I have not added a `--with-gsl` flag to the configure (I do not understand `.m4` wizardry yet :slightly_smiling_face:), so depending on where it is installed for users, they may need to add the location of the `libopenblas.so.0` shared library to their `LD_LIBRARY_PATH` environment variable in order to run.