@andreas-wilm 2021-01-06
Hi Bryce, this is super cool! I have a couple of questions: the file size is only a proxy for coverage (`n`). Would you be able to generate the same table with coverage added? To that end you can just compute the average on the output of `samtools depth -a your.bam` . I'll get to testing this on some real date now