@andreas-wilm 2021-01-06
This works perfectly on all high coverage datasets I have access to (all viral samples)! I'll add this as default into to the build process