@andreas-wilm 2021-01-06
Sorry, one question. The returned `poibin_approximation` is a p-value, right? So I guess we should also use Bonferroni correction as further down below in the code:
    if (pvalue * (double)bonf_factor > sig_level) {goto free_and_exit;}

and thus change

    if (poibin_approximation > sig_level + 0.05) {goto free_and_exit;}

accordingly. Correct? Just double checking...