@andreas-wilm 2021-01-06
See gslapprox branch